Changement int en long pour les timestamp. Correction bug dans le cache (lastGlobalUpdate non mis à jour).
git-svn-id: https://svn.itinerennes.fr/android/trunk/ItineRennes@159 3fbbf0ad-bdca-4ea9-bcf4-e3b57373991e

+++ src/fr/itinerennes/business/cache/GeoCacheProvider.java
@@ -259,7 +259,7 @@ public final class GeoCacheProvider extends AbstractService implements GeoExplor
         boolean isExplored = false;
         while (c.moveToNext()) {
 
-            if (DateUtils.isExpired(c.getInt(1), TTL)) {
+            if (DateUtils.isExpired(c.getLong(1), TTL)) {
                 // ttl expired, remove this bbox from the cache
                 if (LOGGER.isDebugEnabled()) {
                     LOGGER.debug("data for bbox[_id={}] is expired, removing it", c.getInt(0));
